将IPCItem结构体的字段类型从usize更改为u64，以提高兼容性。

+++ src/ipc_item.rs
@@ -1,7 +1,7 @@
 #[repr(C)]
 #[derive(Clone, Copy, Debug)]
 pub struct IPCItem {
-    pub sender: usize,
-    pub msg_type: usize,
-    pub data: [usize; 8],
+    pub sender: u64,
+    pub msg_type: u64,
+    pub data: [u64; 8],
 }
